Classes renamed and files moved
